METDEV-1662: Extend componentDefinitions request
Add File 'og_tag_dialog.json' to contain editor-masks to use for the
og:tags.

Add 'vue-form-generator' to og:tag-tab in pageview and fill it with
Object 'ogTags' from componentDefinition request to be able to change
og:tags in site-info.

Add 'OG_TAG' constant to PerConstants.

Add another level to the result from componentDefinition request.
Previously it gave a schema back which was directly used to fill the
'vue-form-generator' for site-info. This was changed to give back the
whole Object instead of only Object.model so Object.ogTags is
accessable from pageview aswell.

Adjust apiImpl and savePageProperties to use the new Object.model
instead of model